+++ src/util/pup_toNetwork.h
@@ -30,11 +30,11 @@ typedef CMK_NETWORK_INT4 CMK_POINTER_SIZED_INT;
 #endif
 
 class PUP_toNetwork_sizer : public PUP::er {
-	int nBytes;
+	size_t nBytes;
 	virtual void bytes(void *p,int n,size_t itemSize,PUP::dataType t);
  public:
 	PUP_toNetwork_sizer(void) :PUP::er(IS_SIZING), nBytes(0) {}
-	int size(void) const {return nBytes;}
+	size_t size(void) const {return nBytes;}
 };
 
 class PUP_toNetwork_pack : public PUP::er {
@@ -72,7 +72,7 @@ class PUP_toNetwork_pack : public PUP::er {
 		start=buf=(unsigned char *)dest;
 		CmiAssert(sizeof(void *) == sizeof(CMK_POINTER_SIZED_INT));
 	}
-	inline int size(void) const {return buf-start;}
+	inline size_t size(void) const {return buf-start;}
 };
 
 class PUP_toNetwork_unpack : public PUP::er {
@@ -116,7 +116,7 @@ class PUP_toNetwork_unpack : public PUP::er {
 	PUP_toNetwork_unpack(const void *src) :PUP::er(IS_UNPACKING) {
 		start=buf=(const unsigned char *)src;
 	}
-	inline int size(void) const {return buf-start;}
+	inline size_t size(void) const {return buf-start;}
 };
 
 #endif
